Update autoplay system to account for new track URIs
