+++ mopidy_youtube/backend.py
@@ -8,6 +8,7 @@ from mopidy import backend, httpclient
 from mopidy.models import Album, Artist, SearchResult, Track
 from mopidy_youtube import Extension, logger, youtube
 from mopidy_youtube.apis import youtube_api, youtube_bs4api
+from mopidy_youtube.data import format_video_uri, format_playlist_uri, format_channel_uri, extract_video_id, extract_playlist_id, extract_channel_id
 
 
 """
@@ -18,27 +19,6 @@ A typical interaction:
 step 1 requires only 2 API calls. Data for the next steps are loaded in the
 background, so steps 2/3 are usually instantaneous.
 """
-
-
-uri_video_regex = re.compile("^(?:youtube|yt):video:(?P<videoid>.+)$")
-uri_playlist_regex = re.compile("^(?:youtube|yt):playlist:(?P<playlistid>.+)$")
-uri_channel_regex = re.compile("^(?:youtube|yt):channel:(?P<channelid>.+)$")
-
-old_uri_video_regex = re.compile(r"^(?:youtube|yt):video/(?:.+)\.(?P<videoid>.+)$")
-old_uri_playlist_regex = re.compile(r"^(?:youtube|yt):playlist/(?:.+)\.(?P<playlistid>.+)$")
-old_uri_channel_regex = re.compile(r"^(?:youtube|yt):channel/(?:.+)\.(?P<channelid>.+)$")
-
-
-def format_video_uri(video: youtube.Video) -> str:
-    return f"youtube:video:{video.id}"
-
-
-def format_playlist_uri(playlist: youtube.Playlist) -> str:
-    return f"youtube:playlist:{playlist.id}"
-
-
-def format_channel_uri(channel) -> str:
-    return f"youtube:channel:{channel.id}"
 
 
 def convert_video_to_track(video: youtube.Video, album_name: str, **kwargs) -> Track:
@@ -90,30 +70,6 @@ def convert_playlist_to_album(playlist: youtube.Playlist) -> Album:
         artists=[Artist(name=f"YouTube Playlist ({playlist.video_count.get()} videos)")],
         uri=format_playlist_uri(playlist),
     )
-
-
-def extract_video_id(uri) -> str:
-    for regex in (uri_video_regex, old_uri_video_regex):
-        match = regex.match(uri)
-        if match:
-            return match.group("videoid")
-    return ""
-
-
-def extract_playlist_id(uri) -> str:
-    for regex in (uri_playlist_regex, old_uri_playlist_regex):
-        match = regex.match(uri)
-        if match:
-            return match.group("playlistid")
-    return ""
-
-
-def extract_channel_id(uri) -> str:
-    for regex in (uri_channel_regex, old_uri_channel_regex):
-        match = regex.match(uri)
-        if match:
-            return match.group("channelid")
-    return ""
 
 
 class YouTubeBackend(pykka.ThreadingActor, backend.Backend):
